+++ src/main/java/br/com/zalf/prolog/webservice/gente/quiz/quiz/QuizDaoImpl.java
@@ -1,13 +1,13 @@
 package br.com.zalf.prolog.webservice.gente.quiz.quiz;
 
-import br.com.zalf.prolog.webservice.database.DatabaseConnection;
-import br.com.zalf.prolog.webservice.TimeZoneManager;
 import br.com.zalf.prolog.webservice.commons.questoes.Alternativa;
+import br.com.zalf.prolog.webservice.database.DatabaseConnection;
+import br.com.zalf.prolog.webservice.gente.quiz.modelo.QuizModeloConverter;
 import br.com.zalf.prolog.webservice.gente.quiz.quiz.model.AlternativaEscolhaQuiz;
 import br.com.zalf.prolog.webservice.gente.quiz.quiz.model.AlternativaOrdenamentoQuiz;
 import br.com.zalf.prolog.webservice.gente.quiz.quiz.model.PerguntaQuiz;
 import br.com.zalf.prolog.webservice.gente.quiz.quiz.model.Quiz;
-import br.com.zalf.prolog.webservice.gente.quiz.modelo.QuizModeloConverter;
+import org.jetbrains.annotations.NotNull;
 
 import java.sql.*;
 import java.time.Clock;
@@ -26,7 +26,7 @@ public class QuizDaoImpl extends DatabaseConnection implements QuizDao {
     }
 
     @Override
-    public boolean insert(Quiz quiz) throws SQLException {
+    public boolean insert(@NotNull final Quiz quiz) throws Throwable {
         Connection conn = null;
         PreparedStatement stmt = null;
         ResultSet rSet = null;
@@ -54,13 +54,16 @@ public class QuizDaoImpl extends DatabaseConnection implements QuizDao {
             conn.rollback();
             return false;
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
         return true;
     }
 
+    @NotNull
     @Override
-    public List<Quiz> getRealizadosByColaborador(Long cpf, int limit, int offset) throws SQLException {
+    public List<Quiz> getRealizadosByColaborador(@NotNull final Long cpf,
+                                                 final int limit,
+                                                 final int offset) throws Throwable {
         Connection conn = null;
         PreparedStatement stmt = null;
         ResultSet rSet = null;
@@ -70,7 +73,7 @@ public class QuizDaoImpl extends DatabaseConnection implements QuizDao {
             stmt = conn.prepareStatement("SELECT " +
                     "Q.CODIGO AS CODIGO, " +
                     "Q.COD_MODELO AS COD_MODELO, " +
-                    "Q.DATA_HORA AT TIME ZONE ? AS DATA_HORA, " +
+                    "Q.DATA_HORA AT TIME ZONE TZ_UNIDADE(Q.COD_UNIDADE) AS DATA_HORA, " +
                     "QM.NOME AS NOME, " +
                     "Q.QT_CORRETAS AS QT_CORRETAS, " +
                     "Q.QT_ERRADAS AS QT_ERRADAS, " +
@@ -81,31 +84,30 @@ public class QuizDaoImpl extends DatabaseConnection implements QuizDao {
                     "AND Q.COD_UNIDADE = QM.COD_UNIDADE WHERE CPF_COLABORADOR = ? " +
                     "ORDER BY DATA_HORA DESC " +
                     "LIMIT ? OFFSET ?");
-            stmt.setString(1, TimeZoneManager.getZoneIdForCpf(cpf, conn).getId());
-            stmt.setLong(2, cpf);
-            stmt.setInt(3, limit);
-            stmt.setInt(4, offset);
+            stmt.setLong(1, cpf);
+            stmt.setInt(2, limit);
+            stmt.setInt(3, offset);
             rSet = stmt.executeQuery();
             while (rSet.next()) {
                 quizes.add(createQuiz(rSet));
             }
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
         return quizes;
     }
 
+    @NotNull
     @Override
-    public Quiz getByCod(Long codUnidade, Long codQuiz, Long codModeloQuiz) throws Throwable {
+    public Quiz getByCod(@NotNull final Long codQuiz) throws Throwable {
         Connection conn = null;
         PreparedStatement stmt = null;
         ResultSet rSet = null;
-        Quiz quiz = null;
         try {
             conn = getConnection();
             stmt = conn.prepareStatement("SELECT Q.CODIGO AS CODIGO, " +
                     "Q.COD_MODELO AS COD_MODELO, " +
-                    "Q.DATA_HORA AT TIME ZONE ? AS DATA_HORA, " +
+                    "Q.DATA_HORA AT TIME ZONE TZ_UNIDADE(Q.COD_UNIDADE) AS DATA_HORA, " +
                     "QM.NOME AS NOME, " +
                     "Q.QT_CORRETAS AS QT_CORRETAS, " +
                     "Q.QT_ERRADAS AS QT_ERRADAS, " +
@@ -114,23 +116,22 @@ public class QuizDaoImpl extends DatabaseConnection implements QuizDao {
                     "THEN TRUE ELSE FALSE END) AS APROVADO " +
                     "FROM QUIZ q join quiz_modelo QM ON Q.cod_modelo = QM.codigo\n" +
                     "AND Q.cod_unidade = QM.cod_unidade " +
-                    "WHERE Q.CODIGO = ? and Q.cod_modelo = ? and Q.cod_unidade = ?");
-            stmt.setString(1, TimeZoneManager.getZoneIdForCodUnidade(codUnidade, conn).getId());
-            stmt.setLong(2, codQuiz);
-            stmt.setLong(3, codModeloQuiz);
-            stmt.setLong(4, codUnidade);
+                    "WHERE Q.CODIGO = ?;");
+            stmt.setLong(1, codQuiz);
             rSet = stmt.executeQuery();
             if (rSet.next()) {
-                quiz = createQuiz(rSet);
-                quiz.setPerguntas(getPerguntasByCodQuiz(codQuiz, codUnidade, codModeloQuiz, conn));
+                final Quiz quiz = createQuiz(rSet);
+                quiz.setPerguntas(getPerguntasByCodQuiz(codQuiz, conn));
+                return quiz;
+            } else {
+                throw new SQLException("Nenhum quiz encontrado com o código: " + codQuiz);
             }
         } finally {
-            closeConnection(conn, stmt, rSet);
+            close(conn, stmt, rSet);
         }
-        return quiz;
     }
 
-    private void insertRespostas(Quiz quiz, Connection conn) throws SQLException {
+    private void insertRespostas(@NotNull final Quiz quiz, @NotNull final Connection conn) throws SQLException {
         PreparedStatement stmt = null;
         try {
             stmt = conn.prepareStatement("INSERT INTO quiz_respostas(cod_modelo, cod_unidade, cod_quiz, " +
@@ -157,11 +158,12 @@ public class QuizDaoImpl extends DatabaseConnection implements QuizDao {
                 }
             }
         } finally {
-            closeStatement(stmt);
+            close(stmt);
         }
     }
 
-    private Quiz createQuiz(ResultSet rSet) throws SQLException {
+    @NotNull
+    private Quiz createQuiz(@NotNull final ResultSet rSet) throws SQLException {
         final Quiz quiz = new Quiz();
         quiz.setCodigo(rSet.getLong("CODIGO"));
         quiz.setCodModeloQuiz(rSet.getLong("COD_MODELO"));
@@ -174,58 +176,59 @@ public class QuizDaoImpl extends DatabaseConnection implements QuizDao {
         return quiz;
     }
 
-    private List<PerguntaQuiz> getPerguntasByCodQuiz(Long codQuiz, Long codUnidade, Long codModeloQuiz, Connection
-            conn) throws Throwable {
+    @NotNull
+    private List<PerguntaQuiz> getPerguntasByCodQuiz(@NotNull final Long codQuiz,
+                                                     @NotNull final Connection conn) throws Throwable {
         PreparedStatement stmt = null;
         ResultSet rSet = null;
-        final List<PerguntaQuiz> perguntas = new ArrayList<>();
         try {
-            stmt = conn.prepareStatement("SELECT DISTINCT QP.* FROM quiz_respostas QR\n" +
+            stmt = conn.prepareStatement("SELECT DISTINCT QP.* FROM quiz_respostas QR " +
                     "JOIN quiz_perguntas QP ON QP.cod_unidade = QR.cod_unidade AND QP.CODIGO = QR.cod_pergunta " +
-                    "AND QP.cod_modelo = QR.cod_modelo\n" +
-                    "WHERE QR.cod_unidade = ? AND QR.cod_quiz = ? and QR.cod_modelo = ?\n" +
+                    "AND QP.cod_modelo = QR.cod_modelo " +
+                    "WHERE QR.cod_quiz = ? " +
                     "ORDER BY QP.ordem");
-            stmt.setLong(1, codUnidade);
-            stmt.setLong(2, codQuiz);
-            stmt.setLong(3, codModeloQuiz);
+            stmt.setLong(1, codQuiz);
             rSet = stmt.executeQuery();
+            final List<PerguntaQuiz> perguntas = new ArrayList<>();
             while (rSet.next()) {
                 final PerguntaQuiz pergunta = QuizModeloConverter.createPerguntaQuiz(rSet);
-                pergunta.setAlternativas(getAlternativasPergunta(pergunta.getCodigo(), codUnidade, codQuiz,
-                        codModeloQuiz,
-                        pergunta.getTipo(), conn));
+                pergunta.setAlternativas(getAlternativasPergunta(
+                        codQuiz,
+                        pergunta.getCodigo(),
+                        pergunta.getTipo(),
+                        conn));
                 perguntas.add(pergunta);
             }
+            return perguntas;
         } finally {
-            closeConnection(null, stmt, rSet);
+            close(stmt, rSet);
         }
-        return perguntas;
     }
 
-    private List<Alternativa> getAlternativasPergunta(Long codPergunta, Long codUnidade, Long codQuiz,
-                                                      Long codModeloQuiz, String tipoPergunta, Connection conn)
-            throws Throwable {
+    @NotNull
+    private List<Alternativa> getAlternativasPergunta(@NotNull final Long codQuiz,
+                                                      @NotNull final Long codPergunta,
+                                                      @NotNull final String tipoPergunta,
+                                                      @NotNull final Connection conn) throws Throwable {
         PreparedStatement stmt = null;
         ResultSet rSet = null;
-        final List<Alternativa> alternativas = new ArrayList<>();
         try {
             stmt = conn.prepareStatement("SELECT QAP.*, QR.ordem_selecionada, QR.selecionada FROM " +
                     "quiz_alternativa_pergunta QAP JOIN quiz_respostas QR ON " +
                     "QAP.cod_modelo = QR.cod_modelo AND QAP.cod_unidade = QR.cod_unidade AND QAP.cod_pergunta = QR.cod_pergunta " +
                     "AND QAP.codigo = QR.cod_alternativa " +
-                    "WHERE QR.COD_MODELO = ? AND QR.cod_unidade = ? AND QR.cod_quiz = ? AND QR.cod_pergunta = ? " +
+                    "WHERE QR.cod_quiz = ? AND QR.cod_pergunta = ? " +
                     "ORDER BY QAP.ordem;");
-            stmt.setLong(1, codModeloQuiz);
-            stmt.setLong(2, codUnidade);
-            stmt.setLong(3, codQuiz);
-            stmt.setLong(4, codPergunta);
+            stmt.setLong(1, codQuiz);
+            stmt.setLong(2, codPergunta);
             rSet = stmt.executeQuery();
+            final List<Alternativa> alternativas = new ArrayList<>();
             while (rSet.next()) {
                 alternativas.add(QuizModeloConverter.createAlternativa(rSet, tipoPergunta));
             }
+            return alternativas;
         } finally {
-            closeConnection(null, stmt, rSet);
+            close(stmt, rSet);
         }
-        return alternativas;
     }
 }