+++ src/main/java/br/com/zalf/prolog/webservice/integracao/avacorpavilan/cadastro/PneusVeiculo.java
@@ -1,6 +1,8 @@
 
 package br.com.zalf.prolog.webservice.integracao.avacorpavilan.cadastro;
 
+import br.com.zalf.prolog.webservice.integracao.avacorpavilan.requester.AvacorpAvilanRequestStatus;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
@@ -34,7 +36,7 @@ import javax.xml.bind.annotation.XmlType;
     "mensagem",
     "pneus"
 })
-public class PneusVeiculo {
+public class PneusVeiculo implements AvacorpAvilanRequestStatus {
 
     protected boolean sucesso;
     protected String mensagem;