+++ src/main/java/br/com/zalf/prolog/webservice/integracao/avacorpavilan/checklist/ChecklistsFiltro.java
@@ -1,6 +1,8 @@
 
 package br.com.zalf.prolog.webservice.integracao.avacorpavilan.checklist;
 
+import br.com.zalf.prolog.webservice.integracao.avacorpavilan.requester.AvacorpAvilanRequestStatus;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlType;
@@ -33,7 +35,7 @@ import javax.xml.bind.annotation.XmlType;
     "mensagem",
     "checklists"
 })
-public class ChecklistsFiltro {
+public class ChecklistsFiltro implements AvacorpAvilanRequestStatus {
 
     protected boolean sucesso;
     protected String mensagem;